splitted connection.go to connection.go (interface Connection only) and driver.go
